requirements.txt
pydub
playwright
oss2
qrcode[pil]
python-dotenv
requests
